Organize LocalCatalog store
