+++ libs/fluxc-plugin/src/main/kotlin/org/wordpress/android/fluxc/store/pos/localcatalog/PosCatalogStatusResult.kt
@@ -1,0 +1,6 @@
+package org.wordpress.android.fluxc.store.pos.localcatalog
+
+data class PosCatalogStatusResult(
+    val downloadUrl: String?,
+    val status: String,
+)
