+++ libs/fluxc-plugin/src/main/kotlin/org/wordpress/android/fluxc/store/pos/localcatalog/PosGenerateCatalogResult.kt
@@ -1,0 +1,5 @@
+package org.wordpress.android.fluxc.store.pos.localcatalog
+
+data class PosGenerateCatalogResult(
+    val jobId: String,
+)
