+++ libs/fluxc-plugin/src/main/kotlin/org/wordpress/android/fluxc/store/pos/localcatalog/PosLocalCatalogSyncResult.kt
@@ -1,0 +1,7 @@
+package org.wordpress.android.fluxc.store.pos.localcatalog
+
+data class PosLocalCatalogSyncResult(
+    val syncedCount: Int,
+    val hasMore: Boolean,
+    val nextOffset: Int
+)
