+++ libs/fluxc-plugin/src/main/kotlin/org/wordpress/android/fluxc/store/pos/localcatalog/PosVariationsSyncResult.kt
@@ -1,0 +1,7 @@
+package org.wordpress.android.fluxc.store.pos.localcatalog
+
+data class PosVariationsSyncResult(
+    val syncedCount: Int,
+    val hasMore: Boolean,
+    val nextPage: Int
+)
